Fix more clippy errros and add docs (#52)
* More clippy and docs
* Better names for pipelines

+++ benches/parse.rs
@@ -73,7 +73,7 @@ macro_rules! bench_file {
             });
             c.bench(
                 stringify!($name),
-                b.throughput(|data| Throughput::Bytes(data.len() as u32)),
+                b.throughput(|data| Throughput::Bytes(data.len() as u64)),
             );
         }
     };
